+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - connection test command
 - store author name for version
 - generate file preview
+- Italian translation
 
 ## Changed
 - display local time in history