+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+## 6.1.1
 ## Added
 - connection test command
 - store author name for version